requirements.txt
python-telegram-bot==13.7
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
